renamed intake package to end effector
