+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/blucru/common/subsystems/Robot.java
@@ -6,7 +6,7 @@ import com.qualcomm.robotcore.hardware.VoltageSensor;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.blucru.common.subsystems.boxtube.Pivot;
 import org.firstinspires.ftc.teamcode.blucru.common.subsystems.drivetrain.Drivetrain;
-import org.firstinspires.ftc.teamcode.blucru.common.subsystems.intake.Intake;
+import org.firstinspires.ftc.teamcode.blucru.common.subsystems.endeffector.Intake;
 
 import java.util.ArrayList;
